# Lab2/pzpi-22-3-shpak-denys-lab2/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.BackupDatabaseRequest:
    properties:
      backup_path:
        description: Backup path is the path where the backup will be stored
        type: string
    type: object
  handlers.ClearLogsRequest:
    properties:
      days:
        type: integer
    type: object
...
  handlers.RestoreDatabaseRequest:
    properties:
      backup_path:
        description: Backup path is the path where the backup is stored
        type: string
    type: object
...
  handlers.UpdateSystemConfigsRequest:
    properties:
      timeout_sec:
        type: integer
      token_ttl:
        type: integer
    type: object
...
  handlers.UpdateUserRoleRequest:
    properties:
      roleId:
        type: integer
      userId:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: API Specification
  version: "1.0"
paths:
  /admin/backup:
    post:
      consumes:
      - application/json
      description: Backup the database in CSV format per table
      parameters:
      - description: Backup directory path
        in: body
        name: BackupDatabaseRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.BackupDatabaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Backup created
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Backup the database
      tags:
      - admin
  /admin/change-role:
    post:
      consumes:
      - application/json
      description: Update user role
      parameters:
      - description: UpdateUserRoleRequest
        in: body
        name: UpdateUserRoleRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRoleRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update user role
      tags:
      - admin
  /admin/clear-logs:
    post:
      consumes:
      - application/json
      description: Clear logs older than a specified number of days
      parameters:
      - description: ClearLogsRequest
        in: body
        name: ClearLogsRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.ClearLogsRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Clear logs
      tags:
      - admin
  /admin/db-status:
    get:
      consumes:
      - application/json
      description: Get database status
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get database status
      tags:
      - admin
...
  /admin/logs:
    get:
      consumes:
      - application/json
      description: Get server logs
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get server logs
      tags:
      - admin
  /admin/optimize:
    post:
      consumes:
      - application/json
      description: Optimize database
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Optimize database
      tags:
      - admin
  /admin/restore:
    post:
      consumes:
      - application/json
      description: Restore the database from encrypted backup files
      parameters:
      - description: Backup directory path
        in: body
        name: RestoreDatabaseRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.RestoreDatabaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Database restored
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Restore the database from backup
      tags:
      - admin
  /admin/system-configs:
    get:
      consumes:
      - application/json
      description: Get system configs
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get system configs
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update system configs
      parameters:
      - description: UpdateSystemConfigsRequest
        in: body
        name: UpdateSystemConfigsRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateSystemConfigsRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update system configs
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: 'Provide the Bearer token in the format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"